requirements.txt
langchain>=0.3.16,<0.4.0
langchain-openai>=0.3.3,<0.4.0
langchain-community>=0.3.16,<0.4.0
langchain-core>=0.3.33,<0.4.0
langchain-fmp-data>=0.1.0
openai>=1.58.1
streamlit>=1.32.2
beautifulsoup4==4.12.3
lxml==5.1.0
chromadb>=0.4.24
tiktoken>=0.7.0
yfinance==0.2.50
python-dotenv>=1.0.0